+++ context/doc/02-Architecture-WeeklyWeb(开源版科技周刊阅读平台).html
@@ -140,8 +140,8 @@
     <td>Node.js 访问 SQLite 的数据库驱动，用于保存少量用户态数据。</td>
   </tr>
   <tr>
-    <td>pegefind</td>
-    <td>前端或服务端轻量全文搜索，基于静态生成阶段产出的索引文件查询。</td>
+    <td>Fuse.js</td>
+    <td>前端或服务端轻量模糊搜索，基于静态生成阶段产出的 JSON 索引文件查询，支持拼写容错和字段加权。</td>
   </tr>
   <tr>
     <td>Vite</td>
@@ -180,7 +180,7 @@ web 内容生成脚本读取 ../docs
   ↓
 解析期号、标题、日期、封面图、目录、章节、正文、外链
   ↓
-生成静态 HTML、JSON 数据与 pegefind 搜索索引
+生成静态 HTML、JSON 数据与 Fuse.js 搜索索引
   ↓
 Next.js 构建或运行时读取静态产物</pre>
 
@@ -196,7 +196,7 @@ Next.js 构建或运行时读取静态产物</pre>
 <h2>6. 数据边界</h2>
 <ul>
   <li>周刊内容数据来自 <code>docs</code> 生成出的静态产物。</li>
-  <li>搜索数据来自静态生成阶段产出的 pegefind 索引。</li>
+  <li>搜索数据来自静态生成阶段产出的 Fuse.js 索引。</li>
   <li>用户收藏、阅读进度、阅读历史、主题偏好等少量用户态数据保存到 SQLite。</li>
   <li>SQLite 通过 <code>better-sqlite3</code> 接入，不引入独立数据库服务。</li>
   <li>运行时不得扫描 <code>docs</code> 或重新解析 Markdown；除用户状态外，不新增内容持久化表。</li>
@@ -207,7 +207,7 @@ Next.js 构建或运行时读取静态产物</pre>
   <li>页面必须采用响应式布局，兼容桌面端、平板端和移动端。</li>
   <li>首页保持简洁，只突出最新一期周刊和搜索入口。</li>
   <li>周刊详情页默认展示最新一期，并支持通过年份、月份、周筛选其他周刊。</li>
-  <li>搜索使用静态生成的 pegefind 索引，对标题、章节和正文片段进行关键词匹配。</li>
+  <li>搜索使用静态生成的 Fuse.js 索引，对标题、章节和正文片段进行关键词匹配。</li>
   <li>收藏、阅读历史等需要登录的能力在用户触发时再展示登录入口。</li>
 </ul>
