+++ web/src/components/SearchBox.jsx
@@ -1,25 +1,11 @@
-'use client';
-
-import {useRouter} from 'next/navigation';
-import {useRef} from 'react';
-
-export function SearchBox() {
-    const router = useRouter();
-    const inputRef = useRef(null);
-
-    function handleSubmit(e) {
-        e.preventDefault();
-        const query = inputRef.current?.value.trim() || '';
-        router.push(`/search${query ? `?q=${encodeURIComponent(query)}` : ''}`);
-    }
-
+export function SearchBox({defaultValue = ''} = {}) {
     return (
-        <form className="search-box search-box--large" role="search" onSubmit={handleSubmit}>
+        <form className="search-box search-box--large" action="/search" role="search">
             <span className="search-box__icon" aria-hidden="true">⌕</span>
             <input
-                ref={inputRef}
                 name="q"
                 type="search"
+                defaultValue={defaultValue}
                 placeholder="搜索 AI、开源、rsync…"
                 aria-label="搜索周刊内容"
             />
